requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.2
djangorestframework==3.16.0
django-filter==25.1
drf-nested-routers>=0.94.2
drf-spectacular>=0.28.0
idna==3.10
mysqlclient==2.2.7
requests==2.32.3
sqlparse==0.5.3
tenacity==9.1.2
tzdata==2025.2
urllib3==2.4.0
python-dotenv==1.1.0